+++ api/src/admin/router.ts
@@ -2,6 +2,7 @@ import { resolve } from 'node:path'
 import { readFile } from 'node:fs/promises'
 import { Router } from 'express'
 import { session } from '@data-fair/lib-express'
+import { getStatus } from './status.ts'
 
 const router = Router()
 export default router
@@ -14,8 +15,13 @@ router.use(async (req, res, next) => {
 
 let info = { version: process.env.NODE_ENV }
 if (process.env.NODE_ENV === 'production') {
-  info = JSON.parse(await readFile(resolve(import.meta.dirname, '../../BUILD.json'), 'utf8'))
+  info = JSON.parse(await readFile(resolve(import.meta.dirname, '../../../BUILD.json'), 'utf8'))
 }
 router.get('/info', (req, res) => {
   res.send(info)
+})
+
+router.get('/status', async (req, res) => {
+  const status = await getStatus(req)
+  res.send(status)
 })
