+++ api/src/app.ts
@@ -1,14 +1,14 @@
 import { resolve } from 'node:path'
 import express from 'express'
 import { session, errorHandler, createSiteMiddleware, createSpaMiddleware } from '@data-fair/lib-express'
-import catalogsRouter from './routers/catalogs.ts'
-import datasetsRouter from './routers/datasets.ts'
-import identitiesRouter from './routers/identities.ts'
-import importsRouter from './routers/imports.ts'
-import pluginsRegistryRouter from './routers/plugins-registry.ts'
-import pluginRouter from './routers/plugins.ts'
-import publicationsRouter from './routers/publications.ts'
-import adminRouter from './admin.ts'
+import catalogsRouter from './catalogs/router.ts'
+import datasetsRouter from './misc/routers/datasets.ts'
+import identitiesRouter from './misc/routers/identities.ts'
+import importsRouter from './imports/router.ts'
+import pluginsRegistryRouter from './plugins/registry/router.ts'
+import pluginRouter from './plugins/router.ts'
+import publicationsRouter from './publications/router.ts'
+import adminRouter from './admin/router.ts'
 import config, { uiConfig } from '#config'
 
 export const app = express()
