+++ api/src/catalogs/router.ts
@@ -1,85 +1,20 @@
-import type { SessionStateAuthenticated } from '@data-fair/lib-express'
-import type { CatalogPlugin } from '@data-fair/lib-common-types/catalog/index.js'
 import type { Catalog } from '#types'
 import type { CatalogPostReq } from '#doc/catalogs/post-req/index.ts'
 
 import { Router } from 'express'
 import { nanoid } from 'nanoid'
-import eventsQueue from '@data-fair/lib-node/events-queue.js'
 import { assertAccountRole, session, httpError } from '@data-fair/lib-express'
-import { cipher, decipherSecrets } from '@data-fair/catalogs-shared/cipher.ts'
+import { decipherSecrets } from '@data-fair/catalogs-shared/cipher.ts'
 import { resolvedSchema as catalogSchema } from '#types/catalog/index.ts'
 import mongo from '#mongo'
 import config from '#config'
 import findUtils from '#utils/find.ts'
-import { catalogFacets, catalogsWithCounts } from '#utils/catalogsAggregations.ts'
+import { getPlugin } from '../plugins/service.ts'
+import { sendCatalogEvent, prepareCatalog, validateCatalog } from './service.ts'
+import { catalogFacets, catalogsWithCounts } from './aggregations.ts'
 
 const router = Router()
 export default router
-
-/**
- * Helper function to send events related to catalogs
- * @param catalog The catalog object
- * @param actionText The text describing the action (e.g. "a été créé")
- * @param topicAction The action part of the topic key (e.g. "create", "delete")
- * @param sessionState Optional session state for authentication
- */
-const sendCatalogEvent = (
-  catalog: Catalog,
-  actionText: string,
-  topicAction: string,
-  sessionState?: SessionStateAuthenticated
-) => {
-  if (!config.privateEventsUrl && !config.secretKeys.events) return
-
-  eventsQueue.pushEvent({
-    title: `Le catalogue ${catalog.title} ${actionText}`,
-    topic: { key: `catalogs:catalog-${topicAction}:${catalog._id}` },
-    sender: catalog.owner,
-    resource: {
-      type: 'catalog',
-      id: catalog._id,
-      title: catalog.title,
-    }
-  }, sessionState)
-}
-
-/**
- * Check that a catalog object is valid
- * Check if the plugin exists
- * Check if the config is valid
- */
-const validateCatalog = async (catalog: Partial<Catalog>) => {
-  const validCatalog = (await import('#types/catalog/index.ts')).returnValid(catalog)
-  const plugin: CatalogPlugin = await findUtils.getPlugin(validCatalog.plugin)
-  plugin.assertConfigValid(validCatalog.config)
-  return validCatalog
-}
-
-const prepareCatalog = async (catalog: Catalog) => {
-  const ret: {
-    config?: Catalog['config'],
-    secrets?: Catalog['secrets'],
-    capabilities?: Catalog['capabilities']
-  } = {}
-
-  const plugin = await findUtils.getPlugin(catalog.plugin)
-  const prepareRes = await plugin.prepare({
-    catalogConfig: catalog.config,
-    capabilities: catalog.capabilities,
-    secrets: decipherSecrets(catalog.secrets, config.cipherPassword)
-  })
-
-  if (prepareRes.catalogConfig) ret.config = prepareRes.catalogConfig as Catalog['config']
-  if (prepareRes.capabilities) ret.capabilities = prepareRes.capabilities
-  if (prepareRes.secrets) {
-    ret.secrets = {}
-    for (const key of Object.keys(prepareRes.secrets)) {
-      ret.secrets[key] = cipher(prepareRes.secrets[key], config.cipherPassword)
-    }
-  }
-  return ret
-}
 
 // Get the list of catalogs
 router.get('/', async (req, res) => {
@@ -137,7 +72,7 @@ router.post('/', async (req, res) => {
   catalog.owner = catalog.owner ?? sessionState.account
   assertAccountRole(sessionState, catalog.owner, 'admin')
 
-  const plugin = await findUtils.getPlugin(catalog.plugin)
+  const plugin = await getPlugin(catalog.plugin)
   catalog.capabilities = plugin.metadata.capabilities
 
   catalog._id = nanoid()
@@ -256,7 +191,7 @@ router.get('/:id/resources', async (req, res) => {
   assertAccountRole(sessionState, catalog.owner, 'admin')
 
   // Execute the plugin function
-  const plugin = await findUtils.getPlugin(catalog.plugin)
+  const plugin = await getPlugin(catalog.plugin)
   if (!plugin.metadata.capabilities.includes('import')) throw httpError(501, 'Plugin does not support listing resources')
   const datasets = await plugin.list({
     catalogConfig: catalog.config,
