+++ api/src/misc/utils/find.ts
@@ -1,10 +1,5 @@
-import type { CatalogPlugin } from '@data-fair/lib-common-types/catalog/index.js'
-
 import { mongoPagination, mongoProjection, mongoSort, type SessionStateAuthenticated } from '@data-fair/lib-express'
 import { httpError } from '@data-fair/lib-utils/http-errors.js'
-import config from '#config'
-import fs from 'fs-extra'
-import path from 'path'
 
 /**
  * Show all if super admin, otherwise filter by owner
@@ -41,28 +36,9 @@ export const query = (reqQuery: Record<string, string>, fieldsMap: Record<string
   return query
 }
 
-// Get the plugin from the plugins directory
-const pluginsDir = path.resolve(config.dataDir, 'plugins')
-fs.ensureDirSync(pluginsDir)
-export const getPlugin = async (pluginId: string): Promise<CatalogPlugin> => {
-  try {
-    if (!pluginId) throw httpError(400, 'Plugin ID is required')
-
-    const pluginJsonPath = path.join(pluginsDir, pluginId, 'plugin.json')
-    const pluginJson = JSON.parse(fs.readFileSync(pluginJsonPath, 'utf8'))
-    const pluginPath = path.join(pluginsDir, pluginId, pluginJson.version, 'index.ts')
-    if (!pluginPath) throw httpError(404, `No version found in plugin ${pluginId}`)
-    return (await import(pluginPath)).default
-  } catch (e: any) {
-    if (e.message.includes('Cannot find module')) throw httpError(404, `Plugin ${pluginId} not found (or error in plugin : ${e.message})`)
-    throw e // Rethrow other errors
-  }
-}
-
 export default {
   query,
   filterPermissions,
-  getPlugin,
   sort: mongoSort,
   pagination: mongoPagination,
   project: mongoProjection
