+++ api/src/plugins/registry/router.ts
