+++ ui/dts/typed-router.d.ts
@@ -24,6 +24,5 @@ declare module 'vue-router/auto-routes' {
     '/catalogs/[id]': RouteRecordInfo<'/catalogs/[id]', '/catalogs/:id', { id: ParamValue<true> }, { id: ParamValue<false> }>,
     '/catalogs/new': RouteRecordInfo<'/catalogs/new', '/catalogs/new', Record<never, never>, Record<never, never>>,
     '/dataset-publications': RouteRecordInfo<'/dataset-publications', '/dataset-publications', Record<never, never>, Record<never, never>>,
-    '/dev': RouteRecordInfo<'/dev', '/dev', Record<never, never>, Record<never, never>>,
   }
 }
