+++ worker/src/worker.ts
@@ -9,8 +9,8 @@ import { emit as wsEmit, init as wsInit } from '@data-fair/lib-node/ws-emitter.j
 import { startObserver, stopObserver, internalError } from '@data-fair/lib-node/observer.js'
 import upgradeScripts from '@data-fair/lib-node/upgrade-scripts.js'
 import { getNextImportDate } from '@data-fair/catalogs-shared/cron.ts'
-import importTask from './utils/import.ts'
-import publicationTask from './utils/publication.ts'
+import importTask from './lib/import.ts'
+import publicationTask from './lib/publication.ts'
 import config from '#config'
 import mongo from '#mongo'
 import locks from '#locks'
